Update percentage calculator
The previous version was inaccurate both thematically and through implementation.

+++ classwork/classwork-05/src/dev/beriashvili/classwork/percentages/Calculator.java
@@ -1,5 +1,5 @@
 package dev.beriashvili.classwork.percentages;
 
 public interface Calculator {
-    String getPercentage(double initialCapital, int yearsPassed, double closingCapital);
+    String getPercentage(double initialCapital, double yearsPassed, double closingCapital);
 }
